테스트 케이스 작성 중 오류 수정

+++ Good_Buy/src/main/java/com/itwillbs/goodbuy/controller/MemberController.java
@@ -488,17 +488,22 @@ public class MemberController {
 		
 		// 회원 탈퇴 성공 시 첨부파일 제거 작업
 		String realPath = getRealPath(session);
+		
 		// 끝에 "resources/upload"가 포함되어 있다면 지우기
 		if (realPath.endsWith("resources\\upload") || realPath.endsWith("resources/upload")) {
 			realPath = realPath.substring(0, realPath.lastIndexOf("resources\\upload"));
 		}
-		Path path = Paths.get(realPath, member.getMem_profile());
-		try {
-			Files.delete(path);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
 		
+		//프로필 사진이 저장되어 있을 경우에만 삭제
+		if (member.getMem_profile() != null && !member.getMem_profile().isEmpty()) {
+			Path path = Paths.get(realPath, member.getMem_profile());
+			System.out.println("삭제하려는 파일 경로: " + path);
+			try {
+				Files.delete(path);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}
 		// 세션 제거
 		session.invalidate();
